ref(preprod): Simplify onboarding and extract shared component (EME-733)
Replace complex guided steps with simpler panel matching Size Analysis
onboarding style. Extract shared UI into PreprodOnboardingPanel component.

Changes:
- Extract shared onboarding UI into preprodOnboardingPanel.tsx
- Simplify preprodBuildsOnboarding.tsx (135 lines → 41 lines)
- Update preprodOnboarding.tsx to use shared component (109 lines → 14 lines)
- Add Flutter and React Native documentation links
- Add illustration to mobile builds onboarding
- Fix theme property access (familyMono → font.family.mono, fontSize.sm → font.size.sm)

Both onboardings now show:
- Heading and description
- 3 bullet points of benefits
- Product docs + platform-specific docs (iOS, Android, Flutter, React Native)
- Illustration on right side (hidden on mobile)
